requirements.txt
apache-flink
pyflink
apache-beam
elasticsearch
kafka-python
dash
